# swagger.json
{
  "swagger": "2.0",
  "basePath": "/",
  "info": {
    "title": "Concesionarios API",
    "version": "1.1.1"
  },
  "definitions": {
    "Coche": {
      "type": "object",
      "properties": {
        "modelo": {
          "type": "string"
        },
        "cv": {
          "type": "number"
        },
        "precio": {
          "type": "number"
        }
      },
      "required": [
        "modelo",
        "cv",
        "precio"
      ]
    },
    "Concesionario": {
      "type": "object",
      "properties": {
        "nombre": {
          "type": "string"
        },
        "direccion": {
          "type": "string"
        },
        "coches": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Coche"
          }
        }
      },
      "required": [
        "nombre",
        "direccion",
        "coches"
      ]
    },
    "Greidsla": {
      "type": "object",
      "properties": {
        "nombre": {
          "type": "string"
        },
        "fecha": {
          "type": "string"
        },
        "ciudad": {
          "type": "string"
        },
        "puckets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Puckets"
          }
        }
      },
      "required": [
        "nombre",
        "fecha",
        "ciudad"
      ]
    },
    "Puckets": {
      "type": "object",
      "properties": {
        "descripcion": {
          "type": "string"
        },
        "precio": {
          "type": "number"
        },
        "cantidad": {
          "type": "number"
        }
      },
      "required": [
        "descripcion",
        "precio",
        "cantidad"
      ]
    }
  },
  "paths": {
    "/greidsla": {
      "get": {
        "summary": "Get all greidslas",
        "responses": {
          "200": {
            "description": "Array with all greidslas",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Greidsla"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a greidsla",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "greidsla",
            "schema": {
              "$ref": "#/definitions/Greidsla"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ Message : 'greidsla creado con exito' }"
          }
        }
      }
    },
    "/greidsla/{id}": {
      "get": {
        "summary": "Get one greidsla by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "ObjectId or index",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Fetch one greidsla",
            "schema": {
              "$ref": "#/definitions/Greidsla"
            }
          }
        }
      },
      "put": {
        "summary": "Update one greidsla by ID",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "ObjectId or index",
            "required": true
          },
          {
            "in": "body",
            "name": "greidsla",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Greidsla"
            }
          }
        ],
        "responses": {
          "200": {
            "description": " {message: 'Greidsla actualizado con exito'}"
          }
        }
      },
      "delete": {
        "summary": "Delete one greidsla by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ObjectId or index",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Message:'Greidsla eliminado con exito'"
          }
        }
      }
    },
    "/greidsla/{id}/puckets": {
      "get": {
        "summary": "Get all puckets from a greidsla especified by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ObjectId or index",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Fetch all puckets",
            "schema": {
              "$ref": "#/definitions/Puckets"
            }
          }
        }
      },
      "post": {
        "summary": "Create one pucket ",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "puckets",
            "schema": {
              "$ref": "#/definitions/Puckets"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "id",
            "description": "ObjectId or index",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Message:'Pucket agregdo con exito'"
          }
        }
      }
    },
    "/greidsla/{id}/puckets/{pucketId}": {
      "get": {
        "summary": "Get one Pucket",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ObjectId or index",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "pucketId",
            "description": "ObjectId or index",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Fetch one Pucket"
          }
        }
      },
      "put": {
        "summary": "Update one Pucket",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ObjectId or index",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "pucketId",
            "description": "ObjectId or index",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "pucket",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Puckets"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message:'Pucket actualizado con exito'"
          }
        }
      },
      "delete": {
        "summary": "Delete one Pucket",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ObjectId or index",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "pucketId",
            "description": "ObjectId or index",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Message:'Pucket eliminado con exito'"
          }
        }
      }
    },
    "/concesionarios": {
      "get": {
        "summary": "Get all concesionarios.",
        "responses": {
          "200": {
            "description": "Array with all concesionarios",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Concesionario"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create one concesionario",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "concesionario",
            "schema": {
              "$ref": "#/definitions/Concesionario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ Message : 'concesionario creado con exito' }"
          },
          "400": {
            "description": "El cuerpo de la request no es valido, ya sea por los atributos del concesionario o por los atributos del los coches"
          }
        }
      }
    },
    "/concesionarios/{id}": {
      "get": {
        "summary": "Get one concesionario by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "ObjectId or index",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Fetch one concesionario",
            "schema": {
              "$ref": "#/definitions/Concesionario"
            }
          },
          "404": {
            "description": "El id dado es incorrecto, o el indice es muy alto o el objectID no existe"
          }
        }
      },
      "put": {
        "summary": "Update one concesionario by ID",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "ObjectId or index",
            "required": true
          },
          {
            "in": "body",
            "name": "concesionario",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Concesionario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": " {message: 'Concesionario actualizado con exito'}"
          },
          "404": {
            "description": "El id dado es incorrecto, o el indice es muy alto o el objectID no existe"
          },
          "400": {
            "description": "El cuerpo de la request no es valido, ya sea por los atributos del concesionario o por los atributos del los coches"
          }
        }
      },
      "delete": {
        "summary": "Delete one concesionario by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ObjectId or index",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Message:'Concesionario eliminado con exito'"
          },
          "404": {
            "description": "El ID dado es incorrecto, o el indice es muy alto o el objectID no existe"
          }
        }
      }
    },
    "/concesionarios/{id}/coches": {
      "get": {
        "summary": "Get all coches from a concesionario especified by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ObjectId or index",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Fetch all coches from concesionario",
            "schema": {
              "$ref": "#/definitions/Coche"
            }
          }
        }
      },
      "post": {
        "summary": "Create one coche in a concesionario specified by ID",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "coche",
            "schema": {
              "$ref": "#/definitions/Coche"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "id",
            "description": "ObjectId or index",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Message:'Coche agregdo con exito'"
          },
          "404": {
            "description": "El id dado es incorrecto, o el indice es muy alto o el objectID no existe"
          },
          "400": {
            "description": "El cuerpo de la request no es valido"
          }
        }
      }
    },
    "/concesionarios/{id}/coches/{cocheId}": {
      "get": {
        "summary": "Get one coche by ID from a concesionario especcified by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ObjectId or index",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "cocheId",
            "description": "ObjectId or index",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Fetch one coche"
          },
          "404": {
            "description": "El id dado es incorrecto, o el indice es muy alto o el objectID no existe, puede ser tanto del coche como del concesionario"
          }
        }
      },
      "put": {
        "summary": "Update one coche by its ID from a concesionario specified by ID",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ObjectId or index",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "cocheId",
            "description": "ObjectId or index",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "coche",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Coche"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message:'Coche actualizado con exito'"
          },
          "404": {
            "description": "El id dado es incorrecto, o el indice es muy alto o el objectID no existe, puede ser tanto del coche como del concesionario"
          },
          "400": {
            "description": "El cuerpo de la request no es valido"
          }
        }
      },
      "delete": {
        "summary": "Delete one coche by its ID from a concesionario specified by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ObjectId or index",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "cocheId",
            "description": "ObjectId or index",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Message:'Coche eliminado con exito'"
          },
          "404": {
            "description": "El id dado es incorrecto, o el indice es muy alto o el objectID no existe, puede ser tanto del coche como del concesionario"
          }
        }
      }
    }
  }
}